+++ lib/src/xray/protocols/tun_config.dart
@@ -4,6 +4,7 @@ part of 'protocol_settings.dart';
 abstract class TunConfig with _$TunConfig implements XrayInboundSettings {
   const factory TunConfig({
     String? name,
+    String? desc,
     @JsonKey(name: 'mtu') int? mtu,
     List<String>? gateway,
     @JsonKey(name: 'dns') List<String>? dns,
@@ -16,12 +17,13 @@ abstract class TunConfig with _$TunConfig implements XrayInboundSettings {
     final map = asJsonMap(json, 'tun inbound');
     return TunConfig(
       name: map['name'] as String?,
+      desc: map['desc'] as String?,
       mtu: map['mtu'] as int?,
       gateway: (map['gateway'] as List?)?.cast<String>(),
       dns: (map['dns'] as List?)?.cast<String>(),
       userLevel: map['userLevel'] as int?,
-      autoSystemRoutingTable: (map['autoSystemRoutingTable'] as List?)
-          ?.cast<String>(),
+      autoSystemRoutingTable:
+          (map['autoSystemRoutingTable'] as List?)?.cast<String>(),
       autoOutboundsInterface: map['autoOutboundsInterface'] as String?,
     );
   }
@@ -30,12 +32,13 @@ abstract class TunConfig with _$TunConfig implements XrayInboundSettings {
 
   @override
   Map<String, dynamic> toJson() => withoutNulls({
-    'name': name,
-    'mtu': mtu,
-    'gateway': gateway,
-    'dns': dns,
-    'userLevel': userLevel,
-    'autoSystemRoutingTable': autoSystemRoutingTable,
-    'autoOutboundsInterface': autoOutboundsInterface,
-  });
+        'name': name,
+        'desc': desc,
+        'mtu': mtu,
+        'gateway': gateway,
+        'dns': dns,
+        'userLevel': userLevel,
+        'autoSystemRoutingTable': autoSystemRoutingTable,
+        'autoOutboundsInterface': autoOutboundsInterface,
+      });
 }